Extract TypedRoute expression processing into a reusable method.
